Create abstract base class State (#1639)
