+++ src/py/flwr/server/app.py
@@ -39,7 +39,7 @@ from flwr.server.grpc_server.grpc_server import (
 )
 from flwr.server.history import History
 from flwr.server.server import Server
-from flwr.server.state.state import DriverState
+from flwr.server.state import InMemoryState, State
 from flwr.server.strategy import FedAvg, Strategy
 
 ADDRESS_DRIVER_API = "[::]:9091"
@@ -218,18 +218,18 @@ def run_server() -> None:
     log(INFO, "Starting Flower server")
     event(EventType.RUN_SERVER_ENTER)
 
-    # Shared DriverState
-    driver_state = DriverState()
+    # Shared State
+    state = InMemoryState()
 
     # Shared DriverClientManager
     driver_client_manager = DriverClientManager(
-        driver_state=driver_state,
+        state=state,
     )
 
     # Start Driver API
     driver_server = _run_driver_api_grpc(
         address=args.driver_api_address,
-        driver_state=driver_state,
+        state=state,
     )
 
     # Start Fleet API
@@ -276,13 +276,13 @@ def run_server() -> None:
 
 def _run_driver_api_grpc(
     address: str,
-    driver_state: DriverState,
+    state: State,
 ) -> grpc.Server:
     """Run Driver API (gRPC, request-response)."""
 
     # Create Driver API gRPC server
     driver_servicer: grpc.Server = DriverServicer(
-        driver_state=driver_state,
+        state=state,
     )
     driver_add_servicer_to_server_fn = add_DriverServicer_to_server
     driver_grpc_server = generic_create_grpc_server(
