+++ src/py/flwr/server/state/__init__.py
@@ -13,3 +13,12 @@
 # limitations under the License.
 # ==============================================================================
 """Flower server state."""
+
+
+from .in_memory_state import InMemoryState
+from .state import State
+
+__all__ = [
+    "InMemoryState",
+    "State",
+]
